backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.4.0
Pillow==11.3.0
python-decouple==3.8
django-filter==24.3
djoser==2.2.3
djangorestframework-simplejwt==5.3.0
channels==4.1.0
channels-redis==4.2.0
razorpay==1.4.2
selenium==4.26.1
webdriver-manager==4.0.2
requests==2.32.3
openai==1.58.1
google-generativeai==0.8.3
boto3==1.35.84
django-storages==1.14.4
gunicorn==23.0.0
whitenoise==6.8.2